🔨🤖 Extract useClearHashOnReturnToTop; spy returns undefined pre-measure
The hash-clearing policy moves out of SidebarTableOfContents into its
own hook alongside the other TOC hooks. useTocScrollSpy now returns
undefined until its first measurement (vs "" which also means "above
the first section"), so the deep-link guard reads off the type instead
of needing a comment about mount-time behavior. No behavior change.
