+++ site/useTocScrollSpy.ts
@@ -6,6 +6,8 @@ import { useState, useEffect } from "react"
  * reader is currently on: the last target whose top edge has scrolled above a
  * fixed activation line near the top of the viewport. Pure function of scroll
  * position — same model as the topic-page top nav, site/blocks/StickyNav.tsx.
+ * `undefined` until the first measurement, "" once measured while above the
+ * first target.
  *
  * Clicking a TOC link needs no special handling: the anchor jump fires a
  * scroll event and the spy recomputes from the landing position.
@@ -15,8 +17,8 @@ import { useState, useEffect } from "react"
 // page chrome that the TOC advances when a section feels visually current.
 const ACTIVATION_OFFSET_PX = 160
 
-export function useTocScrollSpy(ids: string[]): string {
-    const [activeId, setActiveId] = useState("")
+export function useTocScrollSpy(ids: string[]): string | undefined {
+    const [activeId, setActiveId] = useState<string | undefined>(undefined)
 
     useEffect(() => {
         if (ids.length === 0) return
